Got rid of various warnings in Visual Studio.

+++ compositemodel/src/SurfaceModel.C
@@ -902,8 +902,8 @@ shared_ptr<ftSurface> SurfaceModel::fetchAsSharedPtr(ftFaceBase *face) const
 void SurfaceModel::swapFaces(int idx1, int idx2)
 //===========================================================================
 {
-  if (idx1 < 0 || idx1 >= faces_.size() || idx2 < 0 ||
-      idx2 >= faces_.size())
+  if (idx1 < 0 || idx1 >= (int)faces_.size() || idx2 < 0 ||
+      idx2 >= (int)faces_.size())
     return;  // Do nothing
 
   // Swap
